Add polling kind argument to consumer group poll, add default identifier
This commit introduces a new polling kind argument to the consumer
group poll benchmark. Additionally, the default identifier is now
set to the hostname if not provided, ensuring consistent identification
across benchmarks.
