+++ bench/Cargo.toml
@@ -11,6 +11,7 @@ chrono = "0.4.31"
 clap = { version = "4.5.26", features = ["derive"] }
 figlet-rs = "0.1.5"
 futures = "0.3.31"
+hostname = "0.4.0"
 iggy = { path = "../sdk" }
 iggy-benchmark-report = { path = "report" }
 integration = { path = "../integration" }
