+++ bench/src/args/kinds/mod.rs
@@ -1,0 +1,4 @@
+pub mod consumer_group_poll;
+pub mod poll;
+pub mod send;
+pub mod send_and_poll;
