+++ bench/src/args/mod.rs
@@ -1,6 +1,7 @@
 pub mod common;
 pub mod defaults;
 pub mod kind;
+pub mod kinds;
 
 mod examples;
 mod props;
