+++ bench/src/args/props.rs
@@ -1,4 +1,5 @@
 use super::transport::BenchmarkTransportCommand;
+use iggy::messages::poll_messages::PollingKind;
 use integration::test_server::Transport;
 
 pub trait BenchmarkKindProps {
@@ -13,6 +14,7 @@ pub trait BenchmarkKindProps {
     fn disable_parallel_producer_streams(&self) -> bool;
     fn disable_parallel_consumer_streams(&self) -> bool;
     fn transport_command(&self) -> &BenchmarkTransportCommand;
+    fn polling_kind(&self) -> PollingKind;
     fn validate(&self);
     fn inner(&self) -> &dyn BenchmarkKindProps
     where
