+++ bench/src/benchmarks/consumer_group_benchmark.rs
@@ -4,10 +4,7 @@ use crate::{
     benchmarks::{CONSUMER_GROUP_BASE_ID, CONSUMER_GROUP_NAME_PREFIX},
 };
 use async_trait::async_trait;
-use iggy::{
-    client::ConsumerGroupClient, clients::client::IggyClient, error::IggyError,
-    messages::poll_messages::PollingKind,
-};
+use iggy::{client::ConsumerGroupClient, clients::client::IggyClient, error::IggyError};
 use iggy_benchmark_report::benchmark_kind::BenchmarkKind;
 use integration::test_server::{login_root, ClientFactory};
 use std::sync::Arc;
@@ -99,7 +96,7 @@ impl Benchmarkable for ConsumerGroupBenchmark {
                 warmup_time,
                 self.args.sampling_time(),
                 self.args.moving_average_window(),
-                PollingKind::Next,
+                self.args.polling_kind(),
             );
             let future = Box::pin(async move { consumer.run().await });
             futures.as_mut().unwrap().push(future);
